+++ packages/ocean-react/src/Chart/DoughnutChart.tsx
@@ -1,10 +1,47 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import { Chart as ChartJS, ArcElement } from 'chart.js';
 import { Doughnut } from 'react-chartjs-2';
 import classNames from 'classnames';
 
-import { DoughnutChartProps } from './types/DoughnutChart.types';
-import { options } from './utils/chartUtils';
+export type DoughnutChartProps = {
+  /**
+   * Determines the chart title
+   * @requires
+   */
+  title: string;
+  /**
+   * Determines the chart subtitle
+   * @requires
+   */
+  subtitle: string;
+  /**
+   * Determines the center chart value
+   */
+  centerChartValue?: string;
+  /**
+   * Determines the center chart label
+   */
+  centerChartLabel?: string;
+  /**
+   * Determines graphic counts and legends
+   * @requires
+   */
+  data: IDoughnutChartData[];
+  /**
+   * ClassName to overwrite default style
+   * @default null
+   */
+  className?: string;
+} & React.ComponentPropsWithoutRef<'div'>;
+
+export type IDoughnutChartData = {
+  id: string | number;
+  count: number;
+  percentage: number;
+  color: string;
+  descriptionLevel1: string | number | ReactNode;
+  descriptionLevel2?: string | number | ReactNode;
+};
 
 ChartJS.register(ArcElement);
 
@@ -26,16 +63,30 @@ const DoughnutChart: React.FC<DoughnutChartProps> = ({
     ],
   };
 
+  const options = {
+    responsive: true,
+    plugins: {
+      legend: {
+        display: false,
+      },
+    },
+    borderWidth: 0,
+    cutout: '64%',
+  };
+
   return (
     <div className={classNames('ods-chart', className)} {...rest}>
       <div className="ods-chart__title-and-subtitle">
         <div className="ods-typography ods-typography__heading4">{title}</div>
-        <div className="ods-typography ods-typography__description">
-          {subtitle}
-        </div>
+        <div className="ods-typography ods-typography__caption">{subtitle}</div>
       </div>
       <div className="ods-chart__canvas">
-        <Doughnut data={{ ...formatedData }} options={options} />
+        <Doughnut
+          height={180}
+          width={180}
+          data={{ ...formatedData }}
+          options={options}
+        />
         <div className="ods-chart__canvas__center-legend">
           <div className="ods-typography ods-typography__heading2">
             {centerChartValue}
