+++ packages/ocean-react/src/Chart/stories/0 - Introducion.stories.mdx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs';
 
 import DoughnutChart from '../DoughnutChart';
-import { mockedDoughnutData } from '../utils/chartUtils';
+import { mockedDoughnutData } from '../mocks/doughnutChart';
 
 <Meta
   title="Components/Chart/Introducion"
@@ -39,8 +39,8 @@ The doughnut chart is the version of graphic for data collection in a doughnut f
     <DoughnutChart
       title="Title"
       subtitle="Subtitle"
-      centerChartLabel="center chart label"
-      centerChartValue="center chart value"
+      centerChartLabel="center label"
+      centerChartValue="center value"
       data={mockedDoughnutData}
     />
   </div>
@@ -52,8 +52,8 @@ The doughnut chart is the version of graphic for data collection in a doughnut f
 <DoughnutChart
   title="Title"
   subtitle="Subtitle"
-  centerChartLabel="center chart label"
-  centerChartValue="center chart value"
+  centerChartLabel="center label"
+  centerChartValue="center value"
   data={mockedDoughnutData}
 />
 ```
